+++ src/aixd_ara/components/ara_PlotContoursRequest/code.py
@@ -16,11 +16,7 @@ if plot:
     variable_types = get_dataobject_types(session_id())["dataobject_types"]
     request_dict = reformat_request(request, variable_types)
     print request_dict
-    st[cid] = plot_contours_request(session_id(), request_dict, n_samples, output_type) # if output_type interactive: will launch the plotly fig in browser
+    st[cid] = plot_contours_request(session_id(), request_dict, n_samples, "interactive") # will launch the plotly fig in browser
 
 if cid in st.keys():
     print st[cid]
-    #TODO: add error msg here
-    if output_type == "static" and 'imgstr' in st[cid].keys():
-        imgstr = st[cid]['imgstr']
-        img = convert_str_to_bitmap(imgstr, scale)